This is PR #6172 after resolving the conflict on DEPS

original comment from @daniel-story 
-----

This PR adds an option for SPIRV-Tools to use the mimalloc allocator to substantially improve the performance of memory allocation, especially in multithreaded operation. Because of this approach's overwhelming performance advantages in our extensive benchmarking and in production use with real world shaders, this option is enabled by default when building SPIRV-Tools as a dynamic library and when building the standalone executables (but is not enabled when building as a static library in order to avoid unexpected side effects on whatever the static libraries are statically linked to).

In profiling highly multithreaded (i.e., -j %NUMBER_OF_PROCESSORS%) compilation/optimization of complex, real world game shader corpora including some with total compile times in the tens of hours, we found that approximately 30% of the total time was being spent waiting on locks in malloc/free calls from SPIRV-Tools-opt. This in turn led to concerningly low CPU utilization and artificially lengthened compile times beyond what would be reasonable to expect based on comparison with other shader compilation toolchains.

In considering possible solutions, we found that the Vulkan validation layer had encountered similar issues with SPIRV-Tools in the past and had success in using mimalloc to remedy them (https://github.com/KhronosGroup/Vulkan-ValidationLayers/pull/5083).

By following VVL's lead and modifying our shader compiler's (dynamic) SPIRV-Tools-opt library to use mimalloc, we were able to reduce total compilation time of representative workloads by approximately 40%, and achieve close to 100% CPU utilization for the entire duration of compilation. It should be noted that in our environment glslang and SPIRV-Tools-opt are each used as independent dynamic libraries, so using mimalloc in SPIRV-Tools-opt has no direct impact on the behavior or performance of glslang.

The following is a comparison of total wall clock shader compilation time (including both GLSL compilation using glslang and optimization using SPIRV-Tools-opt) of variously sized subsets of a typical real world workload using the default MSVCRT malloc implementation versus using mimalloc.

| Allocator | # Shaders | Time     | # Threads |
|-----------|-----------|----------|-----------|
| default   | 35000     | 3 h 37 m | 32        |
| mimalloc  | 35000     | 1 h 55 m | 32        |
| default   | 3400      | 677 s    | 32        |
| mimalloc  | 3400      | 414 s    | 32        |
| default   | 1700      | 358 s    | 32        |
| mimalloc  | 1700      | 225 s    | 32        |
| default   | 170       | 235 s    | 1         |
| mimalloc  | 170       | 189 s    | 1         |

The first column indicates which malloc implementation is used. The second column indicates the size of the subset of shader variations compiled. The third column indicates the total elapsed wall clock time to compile the specified number of shaders. The times listed are averages of multiple trials. The final column indicates the number of threads used for concurrent compilation (note that the CPU used for testing had 32 logical cores).

In the most representative sample (35,000 variations using all 32 cores), compilation was 1.5 hours (~40%) faster using mimalloc as compared to using the default malloc implementation. There was a similar effect in smaller sample sizes, demonstrating that the performance benefit of using mimalloc is roughly consistent regardless of the number of shaders compiled.

The performance of single threaded compilation was also tested and likewise demonstrated significant advantages to using mimalloc (with a roughly 25% decrease in compilation time).

The following graph shows the results of optimizing 400 shader variations with the default performance recipe (-O) using the standalone spirv-opt executable as is, versus modified to use mimalloc.

![mimalloc](https://github.com/user-attachments/assets/19dcddd2-6bf0-475c-8ae9-c48e228c9690)

The shader variations used fell broadly into three categories: some that optimized almost instantly, some that took around a quarter of a second to optimize, and some that took around two seconds to optimize.

On average, the standalone spirv-opt executable with mimalloc was found to be 25% faster than when using the default malloc implementation. None of the variations with optimization times above the noise were slower to optimize with mimalloc than without. We also observed less variance in optimization time between runs when using mimalloc compared to using the default allocator.

CPU utilization and memory usage were not formally measured, however informal observation showed that the CPU was fully saturated when using mimalloc, and that total memory usage increased slightly (but not meaningfully) -- because the CPU was substantially busier.

While this change was originally motivated by improving the performance SPIRV-Tools-opt specifically, this PR also adds mimalloc support to the rest of SPIRV-Tools including both the dynamic versions of the libraries and the standalone executables since they all stand to benefit similarly from these improvements.

As an alternative to using a canned solution like mimalloc, it would also be possible to follow glslang's lead and implement a bespoke allocator to achieve similar (or perhaps even greater) performance benefits. This kind of approach has been previously suggested as a solution to this class of problems in https://github.com/KhronosGroup/SPIRV-Tools/issues/1947.

A bespoke allocator would have the advantage of being able to realize the same performance improvements in the static versions of the libraries as it would avoid the need to override the general purpose allocation functions (malloc, free, operator new/delete) and would therefore be safe for static linking into other programs. However, this would require the consistent use of custom allocators in all usages of STL containers and operator new/delete in SPIRV-Tools and would therefore constitute a large and invasive change.

Additionally, developing a custom allocator would take considerably more time and effort and may pose unknown additional risks further downstream in the ecosystem. In comparison, mimalloc is a widely trusted and actively maintained general purpose allocator with a strong track record of existing usage in a broad range of applications, and has already been battle tested in conjunction with SPIRV-Tools by VVL.

We believe that regardless of any further improvements that may be possible in the future that the solution in this PR constitutes a worthwhile improvement in its own right.

Since mimalloc is only used when building the dynamic library and in standalone executables and since its entry points are not exported, users can treat it as an internal implementation detail and it can always be removed or replaced in the future if/when a better solution becomes available.